# code/generated/test_service_server/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: avelasco.horcajada@knowmadmood.com
    name: Antonio Velasco Horcajada
  description: "API for authoring immutable test definitions, composing test plans,\
    \ executing validated plans, and publishing read-only projections to an external\
    \ testing viewer."
  title: Test Service
  version: 0.1.1
servers:
- description: Local development server
  url: http://localhost:8080
security:
- bearerAuth: []
...
paths:
...
  /v1/test-sets/{testSetId}:
    get:
      description: Returns an immutable test set version by UUID.
      operationId: getTestSet
      parameters:
      - description: UUID of the immutable test set version.
        explode: false
        in: path
        name: testSetId
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestSet"
          description: Test set
        "400":
          content:
            application/problem+json:
              example:
                type: /problems/bad-request
                status: 400
                title: Bad Request
                detail: The request is invalid.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The request is invalid.
        "401":
          content:
            application/problem+json:
              example:
                type: /problems/unauthorized
                status: 401
                title: Unauthorized
                detail: Authentication is required or the credentials are invalid.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: Authentication is required or failed.
        "403":
          content:
            application/problem+json:
              example:
                type: /problems/forbidden
                status: 403
                title: Forbidden
                detail: The authenticated principal lacks the required permission.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The authenticated principal is not allowed to perform the operation.
        "404":
          content:
            application/problem+json:
              example:
                type: /problems/not-found
                status: 404
                title: Not Found
                detail: The requested resource was not found.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The requested resource was not found.
        "409":
          content:
            application/problem+json:
              example:
                type: /problems/conflict
                status: 409
                title: Conflict
                detail: The resource state conflicts with this operation.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The operation conflicts with the current state of the resource.
        "422":
          content:
            application/problem+json:
              example:
                type: /problems/domain-validation
                status: 422
                title: Unprocessable Entity
                detail: The request violates a domain rule.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The request is syntactically valid but violates a domain rule.
        "500":
          content:
            application/problem+json:
              example:
                type: /problems/internal-server-error
                status: 500
                title: Internal Server Error
                detail: An unexpected server error occurred.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: An unexpected server error occurred.
        "503":
          content:
            application/problem+json:
              example:
                type: /problems/service-unavailable
                status: 503
                title: Service Unavailable
                detail: The service is temporarily unavailable.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The service is temporarily unavailable.
      summary: Get a test set version
      tags:
      - Composition
  /v1/test-sets/{testSetId}/versions:
    post:
      description: Creates a new immutable draft version of a test set.
      operationId: createTestSetVersion
      parameters:
      - description: UUID of the immutable test set version.
        explode: false
        in: path
        name: testSetId
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTestSetRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestSet"
          description: Created
          headers:
            Location:
              description: URI reference of the created resource.
              explode: false
              required: true
              schema:
                format: uri-reference
                type: string
              style: simple
        "400":
          content:
            application/problem+json:
              example:
                type: /problems/bad-request
                status: 400
                title: Bad Request
                detail: The request is invalid.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The request is invalid.
        "401":
          content:
            application/problem+json:
              example:
                type: /problems/unauthorized
                status: 401
                title: Unauthorized
                detail: Authentication is required or the credentials are invalid.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: Authentication is required or failed.
        "403":
          content:
            application/problem+json:
              example:
                type: /problems/forbidden
                status: 403
                title: Forbidden
                detail: The authenticated principal lacks the required permission.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The authenticated principal is not allowed to perform the operation.
        "404":
          content:
            application/problem+json:
              example:
                type: /problems/not-found
                status: 404
                title: Not Found
                detail: The requested resource was not found.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The requested resource was not found.
        "409":
          content:
            application/problem+json:
              example:
                type: /problems/conflict
                status: 409
                title: Conflict
                detail: The resource state conflicts with this operation.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The operation conflicts with the current state of the resource.
        "422":
          content:
            application/problem+json:
              example:
                type: /problems/domain-validation
                status: 422
                title: Unprocessable Entity
                detail: The request violates a domain rule.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The request is syntactically valid but violates a domain rule.
        "500":
          content:
            application/problem+json:
              example:
                type: /problems/internal-server-error
                status: 500
                title: Internal Server Error
                detail: An unexpected server error occurred.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: An unexpected server error occurred.
        "503":
          content:
            application/problem+json:
              example:
                type: /problems/service-unavailable
                status: 503
                title: Service Unavailable
                detail: The service is temporarily unavailable.
              schema:
                $ref: "#/components/schemas/ErrorDetails"
          description: The service is temporarily unavailable.
      summary: Create a test set version
      tags:
      - Composition
...
components:
...
  schemas:
...
    UUID:
      description: Universally unique identifier.
      example: 550e8400-e29b-41d4-a716-446655440000
      format: uuid
      title: UUID
      type: string
...
    TestSet:
      additionalProperties: false
      description: Immutable version of a composed test set.
      example:
        id: 550e8400-e29b-41d4-a716-446655440000
        project:
          key: IAG
          name: AI Gateway
        setKey: checkout-regression
        version: 1
        name: Checkout regression
        description: Regression tests for checkout
        status: DRAFT
        items:
        - 550e8400-e29b-41d4-a716-446655440000
        - 550e8400-e29b-41d4-a716-446655440000
        createdAt: 2026-01-15T10:30:00Z
        createdBy: user@example.com
      properties:
        id:
          description: Universally unique identifier.
          example: 550e8400-e29b-41d4-a716-446655440000
          format: uuid
          title: UUID
          type: string
        project:
          $ref: "#/components/schemas/ProjectReference"
        setKey:
          description: Stable business key of the test set.
          example: checkout-regression
          maxLength: 100
          minLength: 1
          title: setKey
          type: string
        version:
          description: Monotonically increasing version number.
          example: 1
          minimum: 1
          title: version
          type: integer
        name:
          description: Human-readable test set name.
          example: Checkout regression
          maxLength: 255
          minLength: 1
          title: name
          type: string
        description:
          description: Purpose of the test set.
          example: Regression tests for checkout
          title: description
          type: string
        status:
          description: Lifecycle status of this version.
          enum:
          - DRAFT
          - ACTIVE
          - DEPRECATED
          example: DRAFT
          title: status
          type: string
        items:
          description: Ordered test case versions belonging to this test set. Array
            order determines execution order.
          items:
            $ref: "#/components/schemas/UUID"
          minItems: 1
          title: items
          type: array
          uniqueItems: true
        createdAt:
          description: Creation timestamp in UTC.
          example: 2026-01-15T10:30:00Z
          format: date-time
          title: createdAt
          type: string
        createdBy:
          description: Identity that created the test set.
          example: user@example.com
          maxLength: 255
          minLength: 1
          title: createdBy
          type: string
      required:
      - createdAt
      - createdBy
      - id
      - items
      - name
      - project
      - setKey
      - status
      - version
      title: TestSet
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Authentication using a bearer token.
      scheme: bearer
      type: http